Implement an li optimization and remove symbol table from constant eval
Almost all expressions should only use constants, and the use of labels
is mostly limited to instructions like beq and la. In those cases the
complexity of expressions is limited by R2K's module format already
(references can only be added to or subtracted from) so we wouldn't be
calling evaluate in those cases.

As of this commit, addition.asm assembles to the same text that rasm
outputs.